Add CredentialsResponseHandler tests for regular sync
